+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.13.0
-Do not support buffering. It was broken anyways.
+Remove support for buffering. It was broken anyways.
+
+# 0.12.2
+Make local tracing method (ZipkinTracer::TraceClient.local_component_span) returns the result of block
 
 # 0.12.1
 Allow nesting of local tracing spans